Moved signatures code to its own h/c++ file
