+++ AppleWinExpress2019.vcxproj
@@ -92,6 +92,7 @@
     <ClInclude Include="source\SaveState_Structs_common.h" />
     <ClInclude Include="source\SaveState_Structs_v1.h" />
     <ClInclude Include="source\SerialComms.h" />
+    <ClInclude Include="source\Signatures.h" />
     <ClInclude Include="source\SoundCore.h" />
     <ClInclude Include="source\Speaker.h" />
     <ClInclude Include="source\Speech.h" />
@@ -179,6 +180,7 @@
     <ClCompile Include="source\Riff.cpp" />
     <ClCompile Include="source\SaveState.cpp" />
     <ClCompile Include="source\SerialComms.cpp" />
+    <ClCompile Include="source\Signatures.cpp" />
     <ClCompile Include="source\SoundCore.cpp" />
     <ClCompile Include="source\Speaker.cpp" />
     <ClCompile Include="source\Speech.cpp" />
