+++ source/Video.cpp
@@ -739,8 +739,7 @@ void VideoRefreshScreen(uint32_t uRedrawWholeScreenVideoMode /* =0*/, bool bRedr
 		if (g_gamelink.want_mouse) {
 			CMouseInterface* pMouseCard = GetCardMgr().GetMouseCard();
 			int iOOBX, iOOBY;	// out of bounds
-			pMouseCard->SetPositionRel(g_gamelink.input.mouse_dx, g_gamelink.input.mouse_dy, &iOOBX, &iOOBY);
-			LogOutput("SETTING MOUSE POSITION");
+			pMouseCard->SetPositionRel((long)g_gamelink.input.mouse_dx, (long)g_gamelink.input.mouse_dy, &iOOBX, &iOOBY);
 			// Mouse buttons are LEFT, RIGHT, MIDDLE
 			// TODO: Check also for CONTROL and SHIFT here?
 			// Cache old and new
@@ -769,7 +768,7 @@ void VideoRefreshScreen(uint32_t uRedrawWholeScreenVideoMode /* =0*/, bool bRedr
 								0x08, 0x09, 0x51, 0x57, 0x45, 0x52, 0x54, 0x59, 0x55, 0x49, 0x4F, 0x50, 0xDB, 0xDD, 0x0D,
 								0xA2, 0x41, 0x53, 0x44, 0x46, 0x47, 0x48, 0x4A, 0x4B, 0x4C, 0xBA, 0xDE, 0xC0, 0xA0, 0xDC,
 								0x5A, 0x58, 0x43, 0x56, 0x42, 0x4E, 0x4D, 0xBC, 0xBE, 0xBF, 0xA1, 0x6A, 0xA4, 0x20, 0x14,
-								0x70, 0x71, 0x72, 0x73, 0x74, 0x75, 0x76, 0x77, 0x78, 0x79, 0x90, 0x91, 0x24, 0x26, 0x21,
+								0x70, 0x71, 0x72, 0x73, 0x74, 0x75, 0x76, 0x77, 0x78, 0x79, 0x13, 0x91, 0x24, 0x26, 0x21,
 								0x6D, 0x25, 0x0C, 0x27, 0x6B, 0x23, 0x28, 0x22, 0x2D, 0x2E, 0x2C, 0x00, 0xE2, 0x7A, 0x7B,
 								0x0C, 0xEE, 0xF1, 0xEA, 0xF9, 0xF5, 0xF3, 0x00, 0x00, 0xFB, 0x2F, 0x7C, 0x7D, 0x7E, 0x7F,
 								0x80, 0x81, 0x82, 0x83, 0x84, 0x85, 0x86, 0xED, 0x00, 0xE9, 0x00, 0xC1, 0x00, 0x00, 0x87,
@@ -814,7 +813,7 @@ void VideoRefreshScreen(uint32_t uRedrawWholeScreenVideoMode /* =0*/, bool bRedr
 						lparam = lparam | 0x01000000;	// set extended
 					PostMessageW(g_hFrameWindow, iKeyState, iVK_Code, lparam);
 #ifdef DEBUG
-					LogOutput("SCANCODE, iVK, LPARAM: %04X, %04X, %04X\n", scancode, iVK_Code, LPARAM);
+					LogOutput("SCANCODE, iVK, LPARAM: %04X, %04X, %04X\n", scancode, iVK_Code, lparam);
 #endif DEBUG
 				}
 			}
